Game Loop
Added score & lives. When the player runs out of lives, the score resets to 0 and lives reset to 3. When all bubbles are popped, new bubbles are spawned. This allows the player to continue playing indefinitely.

BUG -- Player pawn (BP_RBPawn) is not generating hit events with spawned bubbles. Spawned bubbles are otherwise behaving normally and colliding with projectiles correctly.